Refactor eventEditor/hooks to create separate directories
