+++ assets/src/domain/eventEditor/hooks/entityActionsMenu/index.ts
@@ -1,0 +1,3 @@
+export { default as useEntityActionsMenuItems } from './useEntityActionsMenuItems';
+
+export { default as useEntityActionsMenuSubscription } from './useEntityActionsMenuSubscription';
