+++ assets/src/domain/eventEditor/hooks/entityActionsMenu/useEntityActionsMenuItems.tsx
@@ -1,22 +1,22 @@
 import React from 'react';
 
 import { Entity } from '@appServices/apollo/types';
-import { useEntityActions, useEntityActionsManager } from '@appLayout/entityActionsMenu';
+import { useEntityActionsService, useEntityActionsMenu } from '@appLayout/entityActionsMenu';
 
 const useEntityActionsMenuItems = <E extends Entity, T extends string>(
 	entityType: T,
 	entity: E,
 	filterByEntityType = true
 ): Array<React.ReactNode> => {
-	const entityActionsManager = useEntityActionsManager(entityType, entity.id);
-	const { getSubscriptions } = useEntityActions('eventEditor');
+	const entityActionsMenu = useEntityActionsMenu(entityType, entity.id);
+	const { getSubscriptions } = useEntityActionsService('eventEditor');
 
-	const { getMenuItems } = entityActionsManager;
+	const { getMenuItems } = entityActionsMenu;
 
-	const subscriptions = getSubscriptions<E, T>({ entityType: filterByEntityType ? entityType : null });
+	const subscriptions = getSubscriptions({ entityType: filterByEntityType ? entityType : null });
 
 	Object.values(subscriptions).forEach(({ callback }) => {
-		callback({ entityType, entity }, entityActionsManager);
+		callback({ entityType, entity }, entityActionsMenu);
 	});
 
 	const menuItems = getMenuItems();
