+++ assets/src/domain/eventEditor/hooks/entityContext/test/useDatetimeContext.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { renderHook } from '@testing-library/react-hooks';
 
 import useDatetimeContext from '../useDatetimeContext';
-import { DatetimeProvider } from '../../services/context/DatetimeContext';
+import { DatetimeProvider } from '@edtrServices/context/DatetimeContext';
 
 describe('useDatetimeContext', () => {
 	it('checks for empty id when used outside the context', () => {
