+++ assets/src/domain/eventEditor/hooks/entityContext/test/useTicketContext.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { renderHook } from '@testing-library/react-hooks';
 
 import useTicketContext from '../useTicketContext';
-import TicketProvider from '../../services/context/TicketContext';
+import TicketProvider from '@edtrServices/context/TicketContext';
 
 describe('useTicketContext', () => {
 	it('checks for error when used outside the context', () => {
