+++ assets/src/domain/eventEditor/hooks/entityContext/useTicketContext.ts
@@ -1,8 +1,8 @@
 import { useContext } from 'react';
 import invariant from 'invariant';
 
-import { TicketContext } from '../services/context/TicketContext';
-import { EntityContextProps } from '../services/context/types';
+import { TicketContext } from '@edtrServices/context/TicketContext';
+import { EntityContextProps } from '@edtrServices/context/types';
 
 const useTicketContext = (): EntityContextProps => {
 	const value = useContext<EntityContextProps>(TicketContext);
