+++ assets/src/domain/eventEditor/ui/datetimes/hooks/useDatesActionMenuHandler.tsx
@@ -5,10 +5,10 @@ import EditDateButton from '../datesList/actionsMenu/EditDateButton';
 import DeleteDateButton from '../datesList/actionsMenu/DeleteDateButton';
 import AssignTicketsButton from '../datesList/actionsMenu/AssignTicketsButton';
 import { Datetime } from '@edtrServices/apollo/types';
-import { SubscriptionCallback } from '@appLayout/entityActionsMenu';
+import { EntityActionsSubscriptionCb } from '@appLayout/entityActionsMenu';
 import { useStatus, TypeName } from '@appServices/apollo/status';
 
-type DatesSubscriptionCallback = SubscriptionCallback<Datetime, 'datetime'>;
+type DatesSubscriptionCallback = EntityActionsSubscriptionCb<Datetime, 'datetime'>;
 
 const useDatesActionMenuHandler = (): DatesSubscriptionCallback => {
 	return useCallback<DatesSubscriptionCallback>(({ entityType, entity: date }, { registerMenuItem }) => {
