+++ assets/src/domain/shared/services/apollo/queries/index.ts
@@ -1,3 +1,7 @@
+export { default as useUpdateCache } from './useUpdateCache';
+
 export * from './currentUser';
 
 export * from './generalSettings';
+
+export * from './types';
